% Generated by build script
\name{eric_document}
\alias{eric_document}
\title{Get a single ERIC document by ID}
\usage{
eric_document()
}
\arguments{
\item{eric_id}{ERIC document ID (e.g. "EJ1234567" or "ED654321")}
}
\value{tibble with one row: id, title, author, source, publicationdate,}
\description{Get a single ERIC document by ID}
