% Generated by build script
\name{eric_facets}
\alias{eric_facets}
\title{Search ERIC with faceted results}
\usage{
eric_facets()
}
\arguments{
\item{query}{Search query}
\item{rows}{Number of results (default 10)}
\item{facet_field}{Facet field: "subject", "author", "publicationtype",}
}
\value{list with $results (tibble) and $facets (tibble: value, count)}
\description{Search ERIC with faceted results}
