% Generated by build script
\name{ipeds_completions}
\alias{ipeds_completions}
\title{Get IPEDS completions (degrees awarded) data}
\usage{
ipeds_completions()
}
\arguments{
\item{year}{Academic year (e.g., 2021). Default 2021.}
\item{limit}{Maximum records to return. Default 100.}
}
\value{tibble: unitid, year, cipcode, award_level, race, sex, awards}
\description{Get IPEDS completions (degrees awarded) data}
